# 3-technical/3.1-system-foundation/architecture/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AKAIUNSAN Attendance + Payroll API
  version: 1.0.0
  description: |
    REST API for AKAIUNSAN attendance and payroll systems (PRD-EPIC-002, Phase 0).

    **Auth:** Bearer JWT (15-min TTL) + refresh token cookie (30 days).
    **Internal:** `X-Internal-API-Key` header (service-to-service only).
    **Rate limits:** Production API instances enforce a global 100 requests/minute limit; auth endpoints add OTP/challenge abuse controls.
    **Time:** ISO 8601 UTC; VN time (UTC+7) at presentation layer.
    **Money:** VNĐ as string (Decimal); never float.
    **Errors:** `{ error: { code, message, details? } }`

    Full schemas live in this file. Full business rules in
    [Domain Specs](../domain-specs.md). Architecture in
    [System Design](../../design-standards/system-design.md).

    Code conventions in [Coding Standards](../../design-standards/coding-standards.md).

  contact:
    name: AKAIUNSAN Engineering
    email: engineering@ak.local
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Employee login (phone + password)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeLogin' }
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              schema: { type: string }
              description: refreshToken cookie (httpOnly, Secure, SameSite=Strict)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/request-otp:
    post:
      tags: [auth]
      summary: Request OTP via SMS
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, pattern: '^\+84[0-9]{9}$', example: '+84912345678' }
      responses:
        '200': { description: OTP sent (always 200 to prevent enumeration) }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/login-otp:
    post:
      tags: [auth]
      summary: Login with phone + OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp]
              properties:
                phone: { type: string, pattern: '^\+84[0-9]{9}$' }
                otp: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
  /v1/auth/admin-login:
    post:
      tags: [auth]
      summary: Admin login (email + password); always starts mandatory TOTP challenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLogin' }
      responses:
        '200':
          description: TOTP challenge created in an httpOnly SameSite=Strict cookie
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Opaque totpChallenge cookie; no challenge token is returned in JSON
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwoFactorRequired' }
  /v1/auth/verify-2fa:
    post:
      tags: [auth]
      summary: Verify TOTP (admin only)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [totpCode]
              properties:
                totpCode: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token (browser cookie or native request body)
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: Native clients only }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  expiresIn: { type: integer, example: 900 }
                  refreshToken: { type: string, description: Present for native/body-token rotation }
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout (revoke refresh token)
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: Native clients only }
      responses:
        '204': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (15-min TTL)
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Shared secret, rotated quarterly
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    EmployeeLogin:
      type: object
      required: [phone, password]
      properties:
        phone: { type: string, pattern: '^\+84[0-9]{9}$' }
        password: { type: string, minLength: 6 }
    AdminLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, example: 900 }
        refreshToken: { type: string, description: Native employee sessions; browser admin sessions use the cookie }
        user: { $ref: '#/components/schemas/User' }
    TwoFactorRequired:
      type: object
      required: [message, expiresIn]
      properties:
        message: { type: string, example: '2FA verification required' }
        expiresIn: { type: integer, example: 300 }

    # ===== User / Employee =====
    User:
      type: object
      required: [id, phone, role, status]
      properties:
        id: { type: string, format: uuid }
        email: { type: [string, 'null'], format: email }
        phone: { type: string }
        role: { type: string, enum: [employee, supervisor, bo_admin, system_admin] }
        status: { type: string, enum: [active, inactive, suspended] }
        employee: { anyOf: [{ $ref: '#/components/schemas/Employee' }, { type: 'null' }] }
    Employee:
      type: object
      required: [id, userId, employeeCode, fullName, hireDate, baseSalary, salaryType, status]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        employeeCode: { type: string, example: NV001 }
        fullName: { type: string, example: 'Nguyễn Văn A' }
        dateOfBirth: { type: [string, 'null'], format: date }
        hireDate: { type: string, format: date }
        baseSalary: { type: string, description: 'VNĐ as string Decimal, e.g. 10000000.00' }
        salaryType: { type: string, enum: [monthly, hourly] }
        hourlyRate: { type: [string, 'null'] }
        bankAccount: { type: [string, 'null'] }
        bankName: { type: [string, 'null'] }
        idNumber: { type: [string, 'null'] }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        user:
          type: object
          required: [id, phone, role, status]
          properties:
            id: { type: string, format: uuid }
            email: { type: [string, 'null'], format: email }
            phone: { type: string }
            role: { type: string, enum: [employee] }
            status: { type: string, enum: [active, inactive, suspended] }
        temporaryPassword: { type: string, description: Present only on employee creation }
...